Fix crash on undo delete measure
When undoing the measure delete, the tick of the spanners is updated before the tick of all the notes, so Spanner::computeStartElement computes garbage if we call it too soon. This causes the trill cue note to end up in the wrong place and cause chaos.

With this, we avoid recomputing start and end element in Spanner::setProperty, because they are recomputed anyway during layout. Additionally, we do some better management of Chord::startingSpanners and Chord::endingSpanners