Add MIDI output synthesizer
There must be a special synthesizer to handle note MIDI events so that they can be sent to MIDI devices even when using synths such as MuseSamplerWrapper.
The MIDI events are sent through the channels in PlayerHandler because they're going to be sent to devices by the notation module which has access to track information.