Fix crash when deleting last note of melisma in some case
The problem was that while laying out the `LyricsLine`, the `ticks` of the `Lyrics` were being recalculated. That would trigger a re-layout of the `Lyrics`, which might in some cases result in the removal of the `LyricsLine` that was just being laid out. This caused the crash.

The solution is to do the `ticks` calculations while laying out the `Lyrics`. Then, based on that, a `LyricsLine` is created or not, and it won't need to be removed or re-added again during this layout call.