Fix keyboard navigation order for dock panels
Old situation: you tell the docking system a navigation panel for each dock panel, and the docking system will mess with the order of this panel.

Problems: if a dock panel contains multiple navigation panels, you need to chain those other panels to the first one. And if the first panel is not really available, as in the case of the mixer toolbar, you can't tell the docking system about it.

New situation: you tell the docking system about the navigation section directly, instead of the navigation panel. The docking system will give you a `contentNavigationPanelOrderStart` property back, which you can use to correctly set the order of all navigation panels in your dock panel.