Fix #20361: Clicking the end of a text selection does nothing.
MuseScore has an odd feature that allows you to resize text selections in the Notation View with the mouse by dragging on the end and moving it around. This however means that, because of the way that inputs are processed, clicking on the end of a selection must not simply place the cursor there as would be expected. This commit changes that feature to be reserved to only be active while Shift is held down, which makes much more sense. Clicking on the end of the selection now behaves as would be expected, clearing the current selection and placing the cursor at the mouse pointer's location.