Set PopupView as parent of PopupWindow_QQuickView
This is to ensure that the PopupWindow_QQuickView is deleted when the corresponding PopupView is deleted. Otherwise, you might end up in the situation that the PopupView has already been deleted (for example, because it was unloaded by a loader) but the PopupWindow_QQuickView is still alive and still visible, and thus stuck on the screen forever.