Revert "Set PopupView as parent of PopupWindow_QQuickView"
This reverts commit 09200b3cc1d3f8e6b930b9616284f912a22f76f2.
That commit did fix a real problem (and a memory leak), but that problem should not occur anymore because of d18f9b3b43784db8164ee7204d8a6f470d9d68d6, and the commit caused other problems that are not quite trivial (#22640 and #22641). Therefore, we'll revert this commit for now.