Fix iocCtx/AccessibleItem crash when opening score with mixer open
The ioc context for `AccessibleItem`s was not found, because `AcessibleItem`s are children of `AbstractNavigation`s, and `AbstractNavigation`s are instantiated as children of QML items, and only those QML items have a QML engine. So now we need to look not just for the parent of the `AccessibleItem`, but for the parent of the parent. This modifies the `iocCtxForQmlObject` so that we can move up in the QObject parent hierarchy as many steps as necessary.