Remove unnecessary debug log from Score::tick2Segment
It is perfectly fine for this method to return a null result, it just means that there is no segment of the specified type in the specified location (which is in many cases exactly the information we want). This debug log is unnecessary and litters our console with hundreds of unnecessary messages.