Fix unreliable ordering of spanners during system layout
SpannerMap is an std::multimap from ticks to Spanner*, so it is guaranteed to be ordered by tick. However, it seems that SpannerMap::findOverlapping, which we use to find the spanners contained in a given range, does not guarantee that the output is ordered in the same way. From what I can see, the output is almost always correctly ordered, but apparently under the wrong conditions (see #23549) it may lose the original order. In that example, deleting an unrelated measure in the next page my cause it to be reordered again.

Having the spanners correctly ordered was probably irrelevant before, but is now fundamental for the new snapping/alignment business, so I'm just doing an std::sort on it to reorder it. Performance shouldn't be affected because, as said, the ordering is almost always correct already.

Couple of thoughts:
- Do we really need to use an external library (intervaltree) just for this findOverlapping operation?
- At some point we need to seriously look into how we store/represent spanners. I think there are better ways (especially now as we have TimeTickAnchors) to store our spanners which are simpler and more efficient than a multimap.