Disallow Ctrl+Shift+dragging Frames to prevent crashes
SingleLayout and SingleDraw are not implemented for these types. In release builds, TLayout is used as a fallback, but that assumes that the Frame has a `system` which is not the case when Ctrl+Shift+dragging. Isn't questionable how useful that would be anyway for frames, so let's disable it for now.