Fix crash when removing bottom staff that contains bracket with span >1
The removed code did the following:
- Before removing the staff, save its index
- Then the staff is removed
- Then a bracket is re-added at the saved staff index
But when the removed staff was the bottom staff, the saved staff index now points beyond the end of the list of staves, causing a crash.

Of course, this can be fixed by adding some extra check. But we can just as well remove this code entirely, and implement a more wholistic algorithm in `Score::adjustBracketsDel`. That way, `NotationParts::removeParts` can also benefit from it, instead of only `NotationParts::removeStaves`. This will be done in the next commit.