Show AppMenuBar in Qt 6.5+
In Qt 6.5+, AppMenuBar's Component.onCompleted doesn't run until the
component has a non-zero height. This solves the chicken-and-egg problem
by assigning an initial height, and then replacing it with the height of
its contents in the onCompleted handler.

I'm not a Qt expert, so I have no idea whether this is the best, most
correct solution to this problem, but it is *a* solution.