Fix MainToolbar, and AppMenuBar in Qt 6.5+
The problem was that a ListView lazily creates delegates as needed to
actually display them. If the size of the ListView is 0, then it doesn't
need to display anything, so it may not create any of the delegates. If
it doesn't create any delegates, then we can't make it size itself to
fit its contents. By setting a minimum size of 1, we force the ListView
to create a delegate that we can then use to determine the ListView's
actual size.

Fixes: #24097