VST: process sequences separately, not all at once (same as in FluidSynth)
After PR #22791, we started to process all sequences in the same buffer (one process call) and use sampleOffset to separate events.
However, it turns out that most plugins completely ignore sampleOffset for the params queue (IParamValueQueue),
and just use the last added one. So now we won't rely on the offset and just process the sequences in different process calls