Fix corruption on Measure::fillGap
This function gets called during the reading of the file, when Score::sigmap() hasn't been generated yet, which means it will return a defult time signature of 4/4, as opposed to the actual time signature of 6/8 in the example score. That results in wrong gap rests being generated and therefore corruption.

It seems that using Measure::timesig() instead is more reliable (and simpler too).