Fix crash accessing uninitialized properties
When MuseScore is built with GCC 13.3.0 with optimizations, statements like

    return lineVisible() ? Sid::pedalEndText : Sid::pedalRosetteEndText;

in `Pedal::getPropertyStyle` are optimized to

    return Sid::pedalRosetteEndText - _lineVisible

However constructors like `Pedal::Pedal` call `EngravingObject::initElementStyle`
when such properties are not initialized yet, and when `initElementStyle` calls
`getPropertyStyle` it receives a bogus result and may crash when it calls
`setProperty` with a wrong `PropertyValue`.

On my machines this causes MuseScore to crash during startup about 40% of the
times. This change fixes the crash by adding default initialiazers.