Attempt to correct VTest reference SHA when PR branch is outdated
Rather than always comparing to the top of the current master branch, compare to the latest common ancestor between master and PR. See  https://github.com/orgs/community/discussions/39880.

Not relevant for the PR containing this commit: for that PR, the VTests will fail anyway, because of the renaming of the CCache script. But for future PRs from contributors, which might not always be fully rebased, this should prevent unexpected failures because of changes on the master branch.