Refactor nightly builds
The idea is to trigger nightly builds from a separate workflow, instead of using a cron schedule on the nightly builds workflows themselves. For the master branch, this doesn't make a big difference, but for the current release branch, it does: in the old situation, the YAML workflow files from the master branch would be used (even while building the current release branch), while other scripts are used from the current release branch. This causes problems when the master branch and the current release branch diverge so that the YAML workflow files from master are not compatible anymore with the other scripts from the current release branch. In the new situation, the YAML files from the current release branch are used. Now, only the input parameters specified by the new trigger action must match the expectations of the YAML workflow files on the current release branch, but that's easy enough to accomplish.