ARM Linux: Fix run_lupdate step
The run_lupdate script obviously depends on Qt, which is installed only inside the Docker container.

(The tx_pull script does not depend on Qt, and does not depend on the run_lupdate step, so can stay outside the Docker container. In fact, it _needs_ to stay outside, because it uses the x86_64 tx binary.)