layout score before calculating the height
it may be that the first page contains only a text box. Then when calculating the height for the video, we will take this first page and get the height of the text box (smallest rectangle containing all (visible) page elements), which is incorrect, since for the video converter we turn off the display of text boxes a little higher in the code.
We need to first make a layout (apply changes) and then get the first page