Prevent erroneous enabling of redo action when there are no parts
It was just a logic error; correct logic would have been
```c++
if (code != UNDO_ACTION_CODE && code != REDO_ACTION_CODE) {
    if (!masterNotation->hasParts()) { return false; }
}
```

I.e. all actions, except undo/redo, can only be handled when the current notation contains parts (instruments).

However, to make the intention even more clear, and to prevent lookup in m_isEnabledMap while we already know which action code it is because we have just *checked* which action code it is, I opted for the solution of this commit.