Prevent ChordRest being selected when starting new range selection
If you first selected any element other than a Note or ChordRest (for example, a Dynamic), and then Shift+clicked a Note, the Note's parent ChordRest would be selected as a list selection. Instead, now that note itself will be selected as a range selection, just like when Shift+clicking an element while nothing is selected or while there is a list selection of notes.