Allow pasting single element onto range selection
Limitation: it will be pasted only once in the top-left corner of the range. To fix this, we'd need to reuse all the complicated logic from `NotationInteraction::applyPaletteElement`; duplicating it is of course no option, so we'd need to extract that somehow to make it reusable. That is currently out of scope.