Use time tick segments when pasting list of elements
Specifically for text items, dynamics in particular, we are no longer limited to just chordrest segments, but we can and should now paste them onto time tick segments, for better accuracy.

For sticking and especially figured bass, the existing code contains some special logic, that I didn't want to touch right now. In the future, we will probably overhaul this part of the code anyway.