Fix binding loop in AudioResourceControl, and simplify
This finally fixes the long-standing binding loop on the `itemImplicitWidth` binding inside the FlatButton in `titleLoader`. It is still unclear what was precisely the cause; a vague theory might be that asking for `titleLoader.width` inside the FlatButton inside the `titleLoader` might somehow depend on the FlatButton's implicitWidth, which indeed depends on `itemImplicitWidth`, and perhaps StyledTextLabel's `implicitWidth` depends on its `width` because of elision/wrap mode? We'll probably never know, but anyway, it's fixed now.