VTests: use correct reference ref
The problem seems actually the opposite of what https://github.com/musescore/MuseScore/pull/24847/commits/1a8ded4d6c267e7064a3ca22725a16c9c229962c tried to solve. Pull requests builds run on a merge commit of the PR branch and the base branch. That means that the latest changes on the base branch (at the time of pushing) will be considered, even if the PR branch is behind the base branch. Therefore, the reference should simply be the base branch, and not the merge-base (=common ancestor) of the PR branch and the base branch, because that would miss the latest changes from the base branch.