Fix #23090: Added style option to use expression font for dynamics
Which font the dynamic should be displayed in is decided in TextBase::createBlocks method. Additionally, fragments are also assigned a type to denote what type of text it contains. It currently contains a plain dynamic one which is for dynamics in expression text style along with a default one for all the other text.