Fix GH#775: Crash due to missing/bad libvorbisfile.3.dynlib
by separating Mac builds for Intel and Apple silicone and using HomeBrew for one and MacPorts for the other.
Also being more explicit for MacPorts which packages to use and in which order, rather than relying on automatic dependencies.
Removing BUILD_UI_MU4, which doesn't apply to Mu3 at all.