Fix crash because of data race in `AsyncImpl::onCall`
Resolves: https://github.com/musescore/MuseScore/issues/25884

We were calling `c.caller->disconnectAsync` after `c.f->call()`, not protected by any mutex. The problem is that `c.caller` is deleted by another thread, between the mutex-protected section and the `disconnectAsync` call.

A solution is to move the `disconnectAsync` call inside the mutex-protected section. It does not really matter that it happens before `c.f->call()` this way; one could even argue it's better this way, because it belongs together with `m_calls.erase(it)`. This way, we can be certain that `c.caller` has not been deleted by another thread iff it is found in `m_calls`: when `c.caller` is being deleted (see `~Asyncable`), it calls its `disconnectAll()` method, which calls `AsyncImpl::disconnectAsync`, which uses the mutex and removes the `Asyncable` from `m_calls`.

A remaining potential problem now, is that `c.caller` may be deleted during `c.f->call()`, which may be problematic inside `c.f->call()`. I'm not sure if it is the responsibility of `AsyncImpl` to mitigate that. If it is, that will either be challenging, or require the use of `recursive_mutex`.