fixed calculation of chord symbol duration
The actual distance between the current chord symbol and the next one may be shorter
than the duration of the current chord symbol segment. This situation occurs, for example,
when we have a measure with a whole rest, and one chord symbol is added to the whole rest segment
and the others are added to the time tick segments. Thus, the whole rest segment lasts for the measure duration,
while the time tick segments last only part of it