Fix crash when adding augmentation dot to list selection
Crashed inside `moveToNextInputPos` because the InputState's `m_segment` is nullptr (and the reason that it's nullptr, is that `selection.startSegment()` is nullptr in `InputState::update`, because the selection is not a range selection). Solution: simply don't call `moveToNextInputPos`, as that doesn't really make sense anyway when not in note input mode.