Fix context of text editing shortcuts
Resolves: https://github.com/musescore/MuseScore/issues/26894

`text-b` and `append-measure` have the same shortcut and were in the same context. Which one would get called, would be determined by the order in which they appear in the user's shortcuts.xml file, which is not really defined. The solution is to put the `text-b` action in the more specific `CTX_NOTATION_TEXT_EDITING`. They would be limited to text editing mode anyway, because of the `&Controller::isEditingText` condition specified while registering the action handler in NotationActionController.