Fix #26963: toolbars move to the right
When restoring deafault layout the status bar, playback toolbar,
"Parts" and "Mixer" buttons move to the right.
This would happen because of the way the currente page was beeing
handled on the restoreDefaultLayout() function. To fix this,
when iterating through the pages if the page is the current
page instead of "cleaning" the state, we restore the page state.