Fix #27188 Fixed Swing and Straight markings format
Swing and Straight markings werent formatted
correctly in the Tempo and Masters palette
like the rest of the elements in the section.
This occured because real Tempo texts
automatically take the format because of the
textStyleType, but since Swing and Straight
elements are simple SystemTexts I add to call
initTextStyleType so it could load the
corresponding text style values.
Also removed the call of the function
setTextStyleType,since it is already called in
initTextStyleType.