Update saxophone ranges in templates
Update the saxophone ranges in the templates to match the ranges in
`instruments.xml`, as fixed in `03bd6187a5 instruments.xml: Fix
saxophone pitch range and hand bell transposition`.

The most important part of this change is decreasing the highest
amateur note on all saxophones by 2 semitones - this removes all
altissimo notes from the amateur range.