Fix #27354: Save floating panel geometry even if hidden in tab during docking
Fix #27354: Save floating panel geometry even if hidden in tab during docking

Previously, `saveLastFloatingGeometry()` would only store geometry if the panel was both floating and visible.

This caused issues for panels that were open but background (tabbed but hidden) at the time of docking—they would not be shown in view menu upon when switching back to score view.

This commit removes the visibility check to ensure that the geometry of all floating panels is preserved, regardless of their tab visibility.