Remove unused fretNumPos property
This looks like the leftover of either an old feature or a half-implemented one. The fret number position is not an individually editable property so the Fret object shouldn't have a private variable for it: it should just return the style value. Incidentally, it also seems that we were writing this property to file but not reading it, so I've removed that too.