Improve copying of Google fonts to the backend docker
Instead of using one `cp` per font (currently 3726) a) use only as many as needed per max. command line len (for me that's just 3) and b) use `mv`, which would result in a lightning fast rename operation if source and destination are in the same file system (but won't harm if not, just decays to a `cp` followed by an `rm`, which the script performs at the end anyhow).