Fix inc/dec duration command when used with dotted notes
Resolves: #28069

The inc/dec duration command used the input state duration as the
initial duration instead of the duration of the selected note.
This started causing issues because the input state duration is changed on
each selection change in NotationNoteInput::updateInputState() when the
"Apply accidentals, [...]" setting is set to "To previous not entered".