midi: add option to disable channel 9 to drum track
Currently, any channel 9 in a MIDI file is automatically added as a drum
instead of a melodic track. Which generally is the correct thing to do,
as this is how MIDI has it specified and standardized.

This new option allows to disable this behaviour though. Which is
especially useful when using DLS (DownLoadable Sound) as DLS has a per
channel bit to specify if it is a drum or melodic track.

Some extracted DLS files from N64 games, extracted via the
N64 Soundbank Tool (https://github.com/jombo23/N64-Tools/tree/master/N64SoundbankTool)
for instance, use a channel 9 and have the drum bit in DLS unset.
So this option would be needed to play such files
correctly.

The ANMP media player (https://github.com/derselbst/ANMP) has a similar option.

According to a disscussion on the FluidSynth issue tracker there
apparently is no way to detect and fix such a conflict between DLS and MIDI
automatically, hence such a manual option seems needed:
https://github.com/FluidSynth/fluidsynth/issues/1576

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>