More handbells symbols
These symbols, unlike the ones I've introduced before, don't attach to a specific chord but can be anchored to any point in time, so they can't be treated as articulations. Instead, they need to be treated as playing techniques (for which we already have a dedicated type that's used for other instruments too), with the only difference that they contain a symbol rather than plain text.