fixed fret box relayout
When inserting a staff, if any harmony or fret diagram has changed, we relayout the fret box.
The problem is that after the insertion, we reset the stack state to fix another issue - see cmdState().reset() in the pasteStaff method.
Let’s add a box relayout again at the start of the layout.