Fix crash in SVG export with tapping
It was calling `TDraw::draw(const Tapping* item, …)` on a `Tapping` with neither a symbol nor Text but only a TappingHalfSlur. In that case, the half slur is drawn via `TDraw::draw(const SlurSegment* item, …)`, and the `Tapping` itself is not drawn at all because its bbox is empty, and elements with empty bbox are normally filtered out for drawing by `BspTree::items`, which checks for `bbox` intersection with the viewport rect, and empty RectF never intersects.

Added a similar bbox check in SVG export, and added a super light safety check in TDraw as well.

Resolves: https://github.com/musescore/MuseScore/issues/29257