Fix stack buffer overflow in accessibleInfo for Ambitus with no tpcs set
When an Ambitus is created, it has by default no tpcs set. In the layout code, this situation is handled by placing the notes on the top and bottom staff line. The `accessibleInfo` method did not take this situation into account, and called `tpc2name`; the "inner" tpc2name method has a stack buffer overflow when an invalid tpc is passed in.