engraving: Cleanup bar line span properties
They used to be integers in the pre-3.0 days, where a staff or bar line
could specify that it spans multiple staves. This then got simplified
to a bool, but this transition wasn't done cleanly. This commit makes up for
that by making sure that we only work with booleans. int values are
converted appropriately when reading pre-3.0 scores.