Legacy TreeModelAdaptor: use QPersistentModelIndex where appropriate
Anywhere where a model index is stored, that should be done using a QPersistentModelIndex, rather than a QModelIndex. So, also inside QML. Make the `data` method return `QPersistentModelIndex`.

With QModelIndex, the combination of pointer and row/column can become outdated. The one in QML did indeed become outdated, and therefore, when passing it to the `expand` method, it wasn't found, so nothing was expanded.
With QPersistentModelIndex, the model keeps a reference to the QPersistentModelIndex, and updates it when rows are moved, so that it has always a valid combination of pointer and row/column.