Refactor AccidentalsLayout without std::list
But with a std::vector range represented by a start and end iterator. Instead of popping from a list, we're now shrinking the vector range, until it's empty. Involves consistent usage of std::remove and std::remove_if, also with reverse iterators.

std::remove[_if] are of course in theory a little bit slower than std::list::remove[_if], because they need to `std::move` some values around. But according to the literature, that cost is easily outweighed by the benefits of std::vector, like cache-friendliness.