Fix haywireness when dragging laissez-vibrer ties
(and since the previous commit also ties and slurs, when no grip was selected)

The problem is that when no grip is selected, `NotationInteraction::drag` calls both `editDrag` and `drag` (the latter via `ElementGroup::drag`). And when calling `editDrag` in this situation, it uses the `EditData` members in different meanings than when dragging grips. Solution: let `editDrag` return without doing anything when `editData.curGrip` is invalid, and let `drag` do the work instead.

Resolves: https://github.com/musescore/MuseScore/issues/29771