Fix add multiple harmony using space merged in undo stack
If you created multiple harmonies by using the Space key to navigate between them, the creation of all these harmonies would be merged in one big undo command.

This is because `endEdit` of the previous harmony would not be called until `NotationInteraction` calls `startEditText` with the next harmony, so the addition of the new harmony happend before the `endEdit` of the previous harmony. Therefore, the creation of this new harmony would be merged into the editing of the previous harmony. This is fixed by calling `endEditText` before creating the next one.