Don't take drastic code path when 'accidentally' laying out full score
There should be a difference between explicitly laying out the entire score from scratch, or just "accidentally" laying out a range that happens to span the entire score (for example, the first system in a single-system score).

This matters especially in `ScorePageViewLayout::prepareScore`, where the `layoutAll` code path is quite drastic.

This is a nice addition to the previous commit, because now the system staff visibility popup won't close anymore after every change when used in single-system scores.