Fix two separate errors in handling header logic
First error is during parts creation, where header segments were ignored. That's not correct because a KeySig might happen to be in a header segment just cause it starts after a system break, but we should still clone it to parts.

Second error is in layout logic, where KeySig segments could be set to disabled when they shouldn't be.