Update obsolete unit test
This test wasn't actually testing what it was supposed to, because

a) The tracks mapping gets reset during Excerpt::createExcerpt, so setting it before does nothing and the test was actually running with no tracks mapping at all, just normal cloned staves.

b) The user can't actually do the operation in that order (set the track mapping first and then create the Excerpt). The user first creates the excerpt and then alters the track mapping using the voice visibility options in the stave popup, so that's what the test should do too.