Remove empty segments without pushing the action to undo stack
The creation of the new segments is done via Measure::getSegment without adding it to the undo stack, so we can't instead add to the undo stack the removal of the empty ones, cause we're making the operation asymmetrical. We should add both or none, and I think none is the simpler option here.