Fix mapping system language en_AU to en_GB
Qt gives us ["en-Latn-AU", "en-AU", ..., "en"]. So it already does simplifying/shortening for us. But we used to do shortening ourselves too: we would shorten "en-Latn-AU", which isn't found, to "en-Latn", which isn't found either, and then to "en", which is then mapped to "en_US". That's undesirable for en-AU.

This commit eliminates the custom shortening, so that we first try "en-Latn-AU", and then "en-AU", which is mapped to "en_GB".

Qt blog post: https://www.qt.io/blog/some-recent-changes-to-choice-of-translation

Resolves: https://github.com/musescore/MuseScore/issues/30217

Review hint: enable "Hide whitespace changes".